[FLINK-15582][tests] Enable batch scheduling tests for both LegacyScheduler and DefaultScheduler
